test: move coordinator and unifi_client tests into their own subpackages
Both areas had multiple test files for a single object (UniFiAlertsCoordinator,
UniFiClient) sitting flat in tests/unit/ with a shared filename prefix. Moved
them into tests/unit/coordinator/ and tests/unit/unifi_client/, mirroring the
existing tests/unit/config_flow/ pattern: each package gets its own
conftest.py holding only the fixtures/helpers relevant to that object, and
test files use relative imports (`.conftest`) instead of bare `from conftest
import`.

This shrinks the global tests/unit/conftest.py from 339 to 145 lines — it now
only holds genuinely cross-cutting helpers (make_hass, make_entry,
patch_setup_entry_collaborators) rather than object-specific test
infrastructure. No behaviour change: same 671 tests, same 98.68% coverage.
